# src/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Provides basic functionality to perform operations against accounts.",
        "title": "Account API",
        "contact": {},
        "license": {},
        "version": "1.0.0"
    },
    "host": "localhost",
    "basePath": "/api/v1",
    "paths": {
        "/accounts/id/{accountId}": {
            "get": {
                "description": "Based off the given account identifier in the URI, the account details are looked up.",
                "produces": [
                    "application/json"
                ],
                "summary": "Exposes functionality to fetch the details of an account once a user has logged in.",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The account identifier returned on login.",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Account"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Account": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}